refactor(vscode): normalize node entrypoint
